fix: down script

+++ scripts/down-docker.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-docker compose down
+FILE="docker-compose.yml"
+
+if [ -f "$FILE" ]; then
+  docker compose down
+fi
